refactor: move serialization utils to own module
